requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyGithub==2.2.0
selenium==4.17.2
webdriver-manager==4.0.1
